Receipts: skip re-send if this jabber account is offline
Given the nature of jabber messages, it’s either now or never, so
delivering stale messages from ages ago is kinda pointless.

I do see that it might be nice to have an API for dropped messages to
allow the user to re-send these messages using a different account or
queue them for later. Patches welcome.